+++ R/monitorDygraph.R
@@ -11,30 +11,26 @@
 #' @return Initiates the interactive dygraph plot in RStudio's 'Viewer' tab.
 #' @examples
 #' \dontrun{
-#' airnow <- airnow_load(20140913, 20141010)
-#' King_Fire <- monitor_subsetByDistance(airnow,
-#'                                       longitude=-120.604,
-#'                                       latitude=38.782,
-#'                                       radius=50)
-#' monitorDygraph(King_Fire, title='KingFire/California/2014', rollPeriod=3)
+#' ca <- airnow_load(2017) %>%
+#'   monitor_subset(tlim=c(20171001,20171101), stateCodes='CA')
+#' Vallejo <- monitor_subset(ca, monitorIDs='060950004_01')
+#' Napa_Fires <- monitor_subsetByDistance(ca,
+#'                                        longitude = Vallejo$meta$longitude,
+#'                                        latitude = Vallejo$meta$latitude,
+#'                                        radius = 50)
+#' monitorDygraph(Napa_Fires, title='Napa Fires in California, Oct. 2017')
 #' } 
 
 monitorDygraph <- function(ws_monitor,
-                           title='title',
-                           ylab='PM2.5 Concentration', 
-                           tlim=NULL,
-                           rollPeriod=1,
-                           showLegend=TRUE) {
+                           title = 'title',
+                           ylab = 'PM2.5 Concentration', 
+                           tlim = NULL,
+                           rollPeriod = 1,
+                           showLegend = TRUE) {
   
   # Sanity check
   if ( monitor_isEmpty(ws_monitor) ) {
     stop("ws_monitor object contains zero monitors")
-  }
-  
-  # Sanity check
-  tzCount <- length(unique(ws_monitor$meta$timezone))
-  if (tzCount > 1) {
-    stop(paste0('Dygraphs cannot be made for data with multiple timezones: ',tzCount,' were found'))
   }
   
   # Convert tlim to POSIXct
@@ -44,9 +40,17 @@ monitorDygraph <- function(ws_monitor,
     dateWindow <- NULL
   }
   
+  # Set timezone
+  tzCount <- length(unique(ws_monitor$meta$timezone))
+  if (tzCount > 1) {
+    warning(paste0(tzCount, ' timezones found. Using UTC time.'))
+    tzone <- 'UTC'
+  } else {
+    tzone <- unique(ws_monitor$meta$timezone)
+  }
+  
   # Simplify access to variables
   datetime <- ws_monitor$data$datetime
-  tzone <- ws_monitor$meta[1,'timezone']
   
   # Create an xts from all data columns except the first which is 'datetime'
   timeseriesData <- xts::xts(ws_monitor$data[,-1],datetime,tzone=tzone)
@@ -66,7 +70,7 @@ monitorDygraph <- function(ws_monitor,
   # Create dygraph
   dygraphs::dygraph(timeseriesData, main=title, ylab=ylab) %>%
     dygraphs::dyOptions(useDataTimezone=TRUE) %>%                       # Always show local time
-    dygraphs::dyLegend(show=show, width=500, labelsSeparateLines=TRUE) %>%
+    dygraphs::dyLegend(show=show, width=250, labelsSeparateLines=TRUE) %>%
     dygraphs::dyRangeSelector(dateWindow=dateWindow) %>%
     dygraphs::dyRoller(rollPeriod=rollPeriod)